# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_dao_model.GpuAnalysis:
    properties:
      command:
        description: 采集到的原始数据
        type: string
      createdAt:
        description: 自动追踪的时间戳
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      historicalMetrics:
        type: string
      id:
        type: integer
      jobID:
        type: integer
      jobName:
        type: string
      llmversion:
        type: string
      namespace:
        type: string
      phase1LLMReason:
        type: string
      phase1Score:
        description: LLM 分析结果
        type: integer
      phase2LLMReason:
        type: string
      phase2Score:
        type: integer
      podName:
        description: 原始 Kubernetes 信息
        type: string
      reviewStatus:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.ReviewStatus'
        description: 管理状态
      userID:
        type: integer
      userName:
        type: string
    type: object
...
  github_com_raids-lab_crater_dao_model.ReviewStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      _: 零值，被忽略
      ReviewStatusConfirmed: 已确认
      ReviewStatusIgnored: 已忽略
      ReviewStatusPending: 待审核
    x-enum-descriptions:
    - 零值，被忽略
    - 待审核
    - 已确认
    - 已忽略
    x-enum-varnames:
    - _
    - ReviewStatusPending
    - ReviewStatusConfirmed
    - ReviewStatusIgnored
...
  github_com_raids-lab_crater_dao_model.UserInfo:
    properties:
      nickname:
        type: string
      username:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Page-internal_handler_vcjob_JobResp:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler_vcjob.JobResp'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.GpuAnalysis'
      msg:
        type: string
    type: object
...
  ? github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_vcjob_JobResp
  : properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Page-internal_handler_vcjob_JobResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_vcjob_JupyterTokenResp:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler_vcjob.JupyterTokenResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_service_vcjob.ForwardType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - _
    - IngressType
    - NodePortType
  github_com_raids-lab_crater_internal_util.VolumeMount:
    properties:
      datasetID:
        type: integer
      mountPath:
        type: string
      subPath:
        type: string
      type:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_util.VolumeType'
    type: object
  github_com_raids-lab_crater_internal_util.VolumeType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - _
    - FileType
    - DataType
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.TriggerJobAnalysisReq:
    properties:
      jobName:
        type: string
    required:
    - jobName
    type: object
...
  internal_handler_vcjob.CreateJupyterReq:
    properties:
      alertEnabled:
        type: boolean
      cpuPinningEnabled:
        type: boolean
      datasetMounts:
        items:
          $ref: '#/definitions/internal_handler_vcjob.DatasetMount'
        type: array
      envs:
        items:
          $ref: '#/definitions/v1.EnvVar'
        type: array
      forwards:
        items:
          $ref: '#/definitions/internal_handler_vcjob.Forward'
        type: array
      image:
        $ref: '#/definitions/internal_handler_vcjob.ImageBaseInfo'
      name:
        type: string
      resource:
        $ref: '#/definitions/v1.ResourceList'
      selectors:
        items:
          $ref: '#/definitions/v1.NodeSelectorRequirement'
        type: array
      template:
        type: string
      volumeMounts:
        items:
          $ref: '#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount'
        type: array
    required:
    - image
    - name
    type: object
  internal_handler_vcjob.DatasetMount:
    properties:
      datasetID:
        type: integer
      mountPath:
        type: string
    type: object
  internal_handler_vcjob.Forward:
    properties:
      name:
        type: string
      port:
        type: integer
      type:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_service_vcjob.ForwardType'
    type: object
  internal_handler_vcjob.ImageBaseInfo:
    properties:
      archs:
        items:
          type: string
        type: array
      imageLink:
        type: string
    type: object
  internal_handler_vcjob.JobResp:
    properties:
      billedPointsTotal:
        type: number
      completedAt:
        type: string
      createdAt:
        type: string
      jobName:
        type: string
      jobType:
        type: string
      locked:
        type: boolean
      lockedTimestamp:
        type: string
      name:
        type: string
      nodes:
        items:
          type: string
        type: array
      owner:
        type: string
      permanentLocked:
        type: boolean
      queue:
        type: string
      resources:
        $ref: '#/definitions/v1.ResourceList'
      startedAt:
        type: string
      status:
        type: string
      userInfo:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.UserInfo'
      waitingToleranceSeconds:
        type: integer
    type: object
  internal_handler_vcjob.JupyterTokenResp:
    properties:
      baseURL:
        type: string
      fullURL:
        type: string
      namespace:
        type: string
      podName:
        type: string
      token:
        type: string
    type: object
...
  resource.Quantity:
    properties:
      Format:
        enum:
        - DecimalExponent
        - BinarySI
        - DecimalSI
        type: string
        x-enum-comments:
          BinarySI: e.g., 12Mi (12 * 2^20)
          DecimalExponent: e.g., 12e6
          DecimalSI: e.g., 12M  (12 * 10^6)
        x-enum-descriptions:
        - e.g., 12e6
        - e.g., 12Mi (12 * 2^20)
        - e.g., 12M  (12 * 10^6)
        x-enum-varnames:
        - DecimalExponent
        - BinarySI
        - DecimalSI
    type: object
  v1.ConfigMapKeySelector:
    properties:
      key:
        description: The key to select.
        type: string
      name:
        description: |-
          Name of the referent.
          This field is effectively required, but due to backwards compatibility is
          allowed to be empty. Instances of this type with an empty value here are
          almost certainly wrong.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
          +optional
          +default=""
          +kubebuilder:default=""
          TODO: Drop `kubebuilder:default` when controller-gen doesn't need it https://github.com/kubernetes-sigs/kubebuilder/issues/3896.
        type: string
      optional:
        description: |-
          Specify whether the ConfigMap or its key must be defined
          +optional
        type: boolean
    type: object
  v1.EnvVar:
    properties:
      name:
        description: Name of the environment variable. Must be a C_IDENTIFIER.
        type: string
      value:
        description: |-
          Variable references $(VAR_NAME) are expanded
          using the previously defined environment variables in the container and
          any service environment variables. If a variable cannot be resolved,
          the reference in the input string will be unchanged. Double $$ are reduced
          to a single $, which allows for escaping the $(VAR_NAME) syntax: i.e.
          "$$(VAR_NAME)" will produce the string literal "$(VAR_NAME)".
          Escaped references will never be expanded, regardless of whether the variable
          exists or not.
          Defaults to "".
          +optional
        type: string
      valueFrom:
        allOf:
        - $ref: '#/definitions/v1.EnvVarSource'
        description: |-
          Source for the environment variable's value. Cannot be used if value is not empty.
          +optional
    type: object
  v1.EnvVarSource:
    properties:
      configMapKeyRef:
        allOf:
        - $ref: '#/definitions/v1.ConfigMapKeySelector'
        description: |-
          Selects a key of a ConfigMap.
          +optional
      fieldRef:
        allOf:
        - $ref: '#/definitions/v1.ObjectFieldSelector'
        description: |-
          Selects a field of the pod: supports metadata.name, metadata.namespace, `metadata.labels['<KEY>']`, `metadata.annotations['<KEY>']`,
          spec.nodeName, spec.serviceAccountName, status.hostIP, status.podIP, status.podIPs.
          +optional
      resourceFieldRef:
        allOf:
        - $ref: '#/definitions/v1.ResourceFieldSelector'
        description: |-
          Selects a resource of the container: only resources limits and requests
          (limits.cpu, limits.memory, limits.ephemeral-storage, requests.cpu, requests.memory and requests.ephemeral-storage) are currently supported.
          +optional
      secretKeyRef:
        allOf:
        - $ref: '#/definitions/v1.SecretKeySelector'
        description: |-
          Selects a key of a secret in the pod's namespace
          +optional
    type: object
...
  v1.NodeSelectorOperator:
    enum:
    - In
    - NotIn
    - Exists
    - DoesNotExist
    - Gt
    - Lt
    type: string
    x-enum-varnames:
    - NodeSelectorOpIn
    - NodeSelectorOpNotIn
    - NodeSelectorOpExists
    - NodeSelectorOpDoesNotExist
    - NodeSelectorOpGt
    - NodeSelectorOpLt
  v1.NodeSelectorRequirement:
    properties:
      key:
        description: The label key that the selector applies to.
        type: string
      operator:
        allOf:
        - $ref: '#/definitions/v1.NodeSelectorOperator'
        description: |-
          Represents a key's relationship to a set of values.
          Valid operators are In, NotIn, Exists, DoesNotExist. Gt, and Lt.
      values:
        description: |-
          An array of string values. If the operator is In or NotIn,
          the values array must be non-empty. If the operator is Exists or DoesNotExist,
          the values array must be empty. If the operator is Gt or Lt, the values
          array must have a single element, which will be interpreted as an integer.
          This array is replaced during a strategic merge patch.
          +optional
          +listType=atomic
        items:
          type: string
        type: array
    type: object
  v1.ObjectFieldSelector:
    properties:
      apiVersion:
        description: |-
          Version of the schema the FieldPath is written in terms of, defaults to "v1".
          +optional
        type: string
      fieldPath:
        description: Path of the field to select in the specified API version.
        type: string
    type: object
...
  v1.ResourceFieldSelector:
    properties:
      containerName:
        description: |-
          Container name: required for volumes, optional for env vars
          +optional
        type: string
      divisor:
        allOf:
        - $ref: '#/definitions/resource.Quantity'
        description: |-
          Specifies the output format of the exposed resources, defaults to "1"
          +optional
      resource:
        description: 'Required: resource to select'
        type: string
    type: object
  v1.ResourceList:
    additionalProperties:
      $ref: '#/definitions/resource.Quantity'
    type: object
  v1.SecretKeySelector:
    properties:
      key:
        description: The key of the secret to select from.  Must be a valid secret
          key.
        type: string
      name:
        description: |-
          Name of the referent.
          This field is effectively required, but due to backwards compatibility is
          allowed to be empty. Instances of this type with an empty value here are
          almost certainly wrong.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
          +optional
          +default=""
          +kubebuilder:default=""
          TODO: Drop `kubebuilder:default` when controller-gen doesn't need it https://github.com/kubernetes-sigs/kubebuilder/issues/3896.
        type: string
      optional:
        description: |-
          Specify whether the Secret or its key must be defined
          +optional
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/gpu-analysis/trigger/job:
    post:
      consumes:
      - application/json
      description: 提供 Job 的 name 来启动一次完整的分析。如果该 Job 已有分析记录，则更新；否则创建新记录。
      parameters:
      - description: Job 信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.TriggerJobAnalysisReq'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功，返回创建或更新后的分析记录
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 分析过程中发生错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 手动触发对单个 Job 的 GPU 分析
      tags:
      - GpuAnalysis
...
  /v1/admin/vcjobs/{name}:
    delete:
      consumes:
      - application/json
      description: 管理员删除用户作业
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Admin delete the job
      tags:
      - VolcanoJob
...
  /v1/vcjobs:
    get:
      consumes:
      - application/json
      description: Get the jobs of the user by client-go
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size, 1-200
        in: query
        name: page_size
        type: integer
      - description: Sort fields
        in: query
        name: sort
        type: string
      - description: Search jobs
        in: query
        name: search
        type: string
      - description: Number of days to look back, -1 for all
        in: query
        name: days
        type: integer
      - collectionFormat: multi
        description: Job types
        in: query
        items:
          type: string
        name: job_type
        type: array
      - collectionFormat: multi
        description: Job statuses
        in: query
        items:
          type: string
        name: status
        type: array
      - description: Node name
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Volcano Job List
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_vcjob_JobResp'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Get the jobs of the user
      tags:
      - VolcanoJob
  /v1/vcjobs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete the job by client-go
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Delete the job
      tags:
      - VolcanoJob
...
  /v1/vcjobs/{name}/snapshot:
    post:
      consumes:
      - application/json
      description: Create nerdctl docker commit to snapshot the job container (supports
        Jupyter and Custom job types)
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_vcjob_JupyterTokenResp'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Create a snapshot of the job container
      tags:
      - VolcanoJob
...
  /v1/vcjobs/{name}/token:
    get:
      consumes:
      - application/json
      description: Get the token of the job by logs
      parameters:
      - description: Job Name
        in: path
        name: jobName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_vcjob_JupyterTokenResp'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Get the ingress base url and jupyter token of the job
      tags:
      - VolcanoJob
...
  /v1/vcjobs/jupyter:
    post:
      consumes:
      - application/json
      description: Create a Jupyter job
      parameters:
      - description: Create Jupyter Job Request
        in: body
        name: CreateJupyterReq
        required: true
        schema:
          $ref: '#/definitions/internal_handler_vcjob.CreateJupyterReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Create a Jupyter job
      tags:
      - VolcanoJob
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"